% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_significance.R
\name{getSig}
\alias{getSig}
\title{Calculate significance statistics}
\usage{
getSig(x, rgraphs)
}
\arguments{
\item{x}{a n x n association matrix returned by \code{\link{makeAssociation}}}

\item{rgraphs}{a list object returned by \code{\link{randomGraphs}}}
}
\value{
A list of significance statistics:
   \item{pl}{A n x n matrix containing the percentage for each cell
     of random association graphs that have values below the value in \code{x}.
     This is essentially a p-value for whether \code{x_ij} is lower than expected}
   \item{pu}{A n x n matrix containing the percentage for each cell
     of random association graphs that have values above the value in \code{x}.
     This is essentially a p-value for whether \code{x_ij} is higher than expected}
   \item{S}{The mean square deviation of the association matrix \code{x}}
   \item{Sp}{The percentage of random association matrices with mean square deviations
     greater than \code{S}. This is essentially a p-value for whether the mean square
     deviation of \code{S} is significantly larger than random, and thus that associations
     overall are occuring significantly more often than random.}
}
\description{
Calculate significance statistics based on a list returned
by \code{\link{randomGraphs}}
}

