% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_statistics.R
\name{gm_mean}
\alias{gm_mean}
\title{a function to calculate the geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{a vector or other object that can be coerced into a vector}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
the geometric mean.
}
\description{
Used internally by \code{\link{wclus}}
}

