% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_significance.R
\name{msdAssoc}
\alias{msdAssoc}
\title{Calculate the mean square deviation of an association matrix}
\usage{
msdAssoc(g, e)
}
\arguments{
\item{g}{an association matrix generated by \code{\link{makeAssociation}}}

\item{e}{the expected association matrix returned by \code{\link{randomGraphs}}
 (The \code{expected} object in the returned list)}
}
\value{
A numeric value equal to the mean square deviation of \code{g}
}
\description{
This calculates the mean square deviation of an association
matrix returned by \code{\link{makeAssociation}}. Used internally by
\code{\link{getSig}}
}

