% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/network_statistics.R
\name{net.density}
\alias{net.density}
\title{A generalized function to calculate network densities}
\usage{
net.density(x, type = "weighted")
}
\arguments{
\item{x}{either a two-mode observational matrix or an association matrix}

\item{type}{One of either "weighted" (default) or "twomode" indicating what kind of
matrix x is}
}
\description{
This function calculates network densities of either two-mode observational matrices
or association matrices.
}

