% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_graphs.R
\name{sampleMatrix}
\alias{sampleMatrix}
\title{Find indices of the matrix that match the pattern for producing random graph}
\usage{
sampleMatrix(x)
}
\arguments{
\item{x}{a m x n matrix or data frame of observational data}
}
\value{
A list containing the rows and columns associated with the matching pattern
}
\description{
Searches for and randomly selects indices in an observation matrix that matches
\code{matrix(c(0,1,1,0),nrow=2)} or \code{matrix(c(1,0,0,1),nrow=2)}. Used internally
by \code{\link{createRandomGraph}} for generating random graphs.
}

