% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_significance.R
\name{test.assoc}
\alias{test.assoc}
\title{Significance test for network statistics of the association matrix}
\usage{
test.assoc(assoc, rgraphs, FUN = net.density, ...)
}
\arguments{
\item{assoc}{A n X n association matrix}

\item{rgraphs}{A list object returned by \code{\link{randomGraphs}}}

\item{FUN}{a function to calculate a network statistic. The function should return a singular
numeric value for which a mean and standard deviation can be calculated.}

\item{...}{Additional arguments to pass to \code{FUN}}
}
\value{
A list containing the following values:
   \item{value}{the value of the test statistic for \code{assoc}}
   \item{random.value}{the mean value of the test statistic for all random graphs in \code{rgraphs}}
   \item{sd.value }{the standard deviation of \code{random.value}}
   \item{p.up}{the percentage of random graphs with values higher than \code{value}}
   \item{p.down}{the percentage of random graphs with values lower than \code{value}}
}
\description{
Test whether a network statistic of an association matrix is
significantly different than the mean of the network statistic
for the random graphs.
}

