% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_statistics.R
\name{wclust}
\alias{wclust}
\title{Weighted clustering}
\usage{
wclust(x, mode = c("athr", "geo", "min", "max"))
}
\arguments{
\item{x}{a n X n association matrix}

\item{mode}{one of "athr" (arithmetic mean), "geo" (geometric mean), "min", or "max" indicating which statistic to use for triplet values}
}
\description{
A function to calculate the weighted clustering of an association matrix
}
\details{
This function calculates a clustering coefficient for a weighted graph by calculating a value
for every triplet of nodes based on the statistic chosen in \code{mode}. The clustering coefficient
is equal to the sum of triplet values for closed triads divided by the sum of all triplet values.
}
\references{
Ospsahl, T. and P. Panzarasa. 2009. "Clustering in weighted network." Social Networks, 31:155-163.
}

