% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_statistics.R
\name{wdegree}
\alias{wdegree}
\title{Calculate the weighted degree of an association matrix}
\usage{
wdegree(x, alpha = 1)
}
\arguments{
\item{x}{a n X n association matrix}

\item{alpha}{The value of the tuning parameter. The value should be a number greater
than or equal to 0. See below for details.}
}
\value{
A named vector of degree scores calculated for each node in the association matrix.
}
\description{
This uses the method described by Opsahl, Agneessens, and Svoretz (2010) to calculate a weighted
degree centrality for each node in an association matrix.
}
\details{
The tuning parameter, \code{alpha}, adjudicates the importance of the number of ties
 a node has versus the strength of those ties in calculating the weighted degree. If \code{alpha} is
 equal to zero, then the resulting degree scores are identical to a calculation of undirected
 degree scores on an unvalued matrix in which any value > 0 counts as a tie. If \code{alpha} is
 equal to 1, then the degree scores are equal to the sum of the values of the ties connected
 to the node. If \code{alpha} is greater than 1, then the number of ties a node has is weighted
 inversely to the degree score, so that the more ties a node has, the lower its degree score will be.
}
\references{
Opsahl, T., F. Agneessens, and J. Svoretz. "Node centrality in weighted networks:
Generalizing degree and shortest paths." Social Networks, 32:245-251.
}

