% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whassocr.R
\docType{package}
\name{whassocr}
\alias{whassocr}
\alias{whassocr-package}
\title{whassocr: An Implementation of Whitehead's Association Analysis}
\description{
This package implements Whitehead's method for analyzing
association data. While Whitehead developed these techniques for animal
associations, this package is agnostic as to the nature of the data, but
assumes data are generally structured as entities in columns and
observations in rows, with cells=c(0,1) indicating whether entities
were observed during the row observation.
}
\section{whassocr functions}{

makeAssociation
}

